% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module.R
\name{angle_deltas}
\alias{angle_deltas}
\title{Todo this text.}
\usage{
angle_deltas(num_clusters, angle_disp)
}
\arguments{
\item{num_clusters}{TODO.}

\item{angle_disp}{TODO.}
}
\value{
TODO.
}
\description{
Todo this text.
}
