% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module.R
\name{clucenters}
\alias{clucenters}
\title{Todo this text.}
\usage{
clucenters(num_clusters, clu_sep, clu_offset)
}
\arguments{
\item{num_clusters}{TODO.}

\item{clu_sep}{TODO.}

\item{clu_offset}{TODO.}
}
\value{
TODO.
}
\description{
Todo this text.
}
