% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clugenr-package.R
\docType{package}
\name{clugenr-package}
\alias{clugenr}
\alias{clugenr-package}
\title{clugenr: Multidimensional Cluster Generation Using Support Lines}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

An implementation of the clugen algorithm for generating multidimensional clusters with arbitrary distributions. Each cluster is supported by a line segment, the position, orientation and length of which guide where the respective points are placed. This package is described in Fachada & de Andrade (2023) \doi{10.1016/j.knosys.2023.110836}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://clugen.github.io/clugenr/}
  \item Report bugs at \url{https://github.com/clugen/clugenr/issues}
}

}
\author{
\strong{Maintainer}: Nuno Fachada \email{faken@fakenmc.com} (\href{https://orcid.org/0000-0002-8487-5837}{ORCID}) [copyright holder]

}
\keyword{internal}
