% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clugenr.R
\docType{package}
\name{clugenr}
\alias{clugenr}
\title{clugenr: Multidimensional cluster generation in R}
\description{
clugenr is an R package for generating multidimensional clusters. Each
cluster is supported by a line segment, the position, orientation and length
of which guide where the respective points are placed.
}
\note{
MIT License
}
\author{
Nuno Fachada
}
