% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{clumerge}
\alias{clumerge}
\title{Merge clusters.}
\usage{
clumerge(..., fields = c("points", "clusters"), clusters_field = "clusters")
}
\arguments{
\item{...}{Data.}

\item{fields}{Fields.}

\item{cluster_field}{Yep.}
}
\value{
The merged data.
}
\description{
To do.
}
\details{
To do.
}
