% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module.R
\name{clupoints_n_1}
\alias{clupoints_n_1}
\title{Todo this text.}
\usage{
clupoints_n_1(projs, lat_disp, line_len, clu_dir, clu_ctr)
}
\arguments{
\item{projs}{TODO.}

\item{lat_disp}{TODO.}

\item{line_len}{TODO.}

\item{clu_dir}{TODO.}

\item{clu_ctr}{TODO.}
}
\value{
TODO. TODO.
}
\description{
Todo this text.
}
