% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module.R
\name{clusizes}
\alias{clusizes}
\title{Todo this text.}
\usage{
clusizes(num_clusters, num_points, allow_empty)
}
\arguments{
\item{num_clusters}{TODO.}

\item{num_points}{TODO.}

\item{allow_empty}{TODO.}
}
\value{
TODO.
}
\description{
Todo this text.
}
\note{
This function is stochastic. For reproducibility set a PRNG seed with
\link{set.seed}.
}
