% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{gdim}
\alias{gdim}
\title{Returns the actual dimensions of the input}
\usage{
gdim(a)
}
\description{
Gets the dimensions of the input \code{a}, returning \code{c(length(a), 1)} if
\code{dim(a) == NULL}.
}
\keyword{internal}
