% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module.R
\name{llengths}
\alias{llengths}
\title{Todo this text.}
\usage{
llengths(num_clusters, llength, llength_disp)
}
\arguments{
\item{num_clusters}{TODO.}

\item{llength}{TODO.}

\item{llength_disp}{TODO.}
}
\value{
TODO.
}
\description{
Todo this text.
}
\note{
This function is stochastic. For reproducibility set a PRNG seed with
\link{set.seed}.
}
