% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{points_on_line}
\alias{points_on_line}
\title{Determine coordinates of points on a line with \code{center} and \code{direction},
based on the distances from the center given in \code{dist_center}.}
\usage{
points_on_line(center, direction, dist_center)
}
\arguments{
\item{center}{Center of the line (n-component vector).}

\item{direction}{Line direction (n-component unit vector).}

\item{dist_center}{Distance of each point to the center of the line
(p-component vector, where p is the number of points).}
}
\value{
Coordinates of points on the specified line (p x n matrix).
}
\description{
Determine coordinates of points on a line with \code{center} and \code{direction},
based on the distances from the center given in \code{dist_center}.
}
