% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{rand_unit_vector}
\alias{rand_unit_vector}
\title{Get a random unit vector with \code{num_dims} components.}
\usage{
rand_unit_vector(num_dims)
}
\arguments{
\item{num_dims}{Number of components in vector (i.e. vector size).}
}
\value{
A random unit vector with \code{num_dims} components.
}
\description{
Get a random unit vector with \code{num_dims} components.
}
\examples{

r <- rand_unit_vector(4)
norm(r, "2")
# 1
}
