% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\docType{data}
\name{tprom}
\alias{tprom}
\title{Generality of types, by increasing order}
\format{
An object of class \code{list} of length 7.
}
\usage{
tprom
}
\description{
Generality of types, by increasing order
}
\keyword{internal}
